Refactor: moved renderer logic to system
to match the ecs pattern we need to keep functionality in the systems and state in the components. 
i'll keep sampling logic in the components "because we only use it to access the state" kappa
next i'll have to move the ressources from the components as well. this is not only to obey the ecs style but has a significant performance inpact as well. the textures in the renderercomponents are highly reusable and don't need to be instantiated with every renderercomponent.
ill move them into the renderersystem for now, but we might have to create some dedicated instance later